Able to edit Student Profile, Slight fixes to StudentHome and CompanyHome to get rid of 'Stuff' subscriptions
User experience is messed up. For this to work, user needs to Add a new profile in the dropdown bar in the top right for anything to show. But works fine after. Will improve at a later date. Also need to fix the design of the profile because it is too basic.